Robust video URL detection algorithm improved!
Video URL detection algorithm now scans all <script> tags and uses
RegEx's to quickly find the video identifier. Then extracts the video
date from the first frame's URL. If and HTTP request to the "estimated"
URL returns 404, then a new HTTP request through a Spanish proxy is
sent in order to figure out the right location of the video. Since this
proxy is unreliable, a retrial system is implemented. Meanwhile, the
user can get status (text) updates of the geo-unblocking process in the
video container (div).